+++ modules/juce_core/native/juce_mac_Strings.mm
@@ -32,7 +32,7 @@ String String::fromCFString (CFStringRef cfString)
     CFIndex bytesNeeded = 0;
     CFStringGetBytes (cfString, range, kCFStringEncodingUTF8, 0, false, nullptr, 0, &bytesNeeded);
 
-    HeapBlock<UInt8> utf8 ((size_t) bytesNeeded + 1);
+    HeapBlock<UInt8> utf8 (bytesNeeded + 1);
     CFStringGetBytes (cfString, range, kCFStringEncodingUTF8, 0, false, utf8, bytesNeeded + 1, nullptr);
 
     return String (CharPointer_UTF8 ((const CharPointer_UTF8::CharType*) utf8.get()),
